+++ Chip8.h
@@ -141,7 +141,7 @@ private:
 	uint16_t opcode;
 
 	std::default_random_engine randGen;
-	std::uniform_int_distribution<uint8_t> randByte;
+	std::uniform_int_distribution<unsigned short> randByte;
 
 	typedef void (Chip8::* Chip8Func)();
 	Chip8Func table[0xF + 1];
